add models

+++ server/src/main/java/ca/fubi/player/repository/RoleRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import ca.fubi.player.models.EnumRole;
 import ca.fubi.player.models.Role;
+import ca.fubi.player.models.enums.EnumRole;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
   Optional<Role> findByName(EnumRole name);
 }
